refactor: organize axios api instance adapter to accept diffente workspaces and add workspace arg to all api methods
